Implement HDR & Tonemapping

+++ CrescentEngine/Rendering/Framebuffer.cpp
@@ -29,7 +29,7 @@ namespace CrescentEngine
 		//Setup Color Buffer.
 		glGenTextures(1, &m_ColorAttachmentID);
 		glBindTexture(GL_TEXTURE_2D, m_ColorAttachmentID);
-		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA8, m_FramebufferWidth, m_FramebufferHeight, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
+		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, m_FramebufferWidth, m_FramebufferHeight, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
